[API] refactor checkin trait, use transaction
